+++ backend/gonzales/static/assets/circle-check-big-C2vTlcqL.js
@@ -1,4 +1,4 @@
-import{c}from"./index-C_tx9rzG.js";/**
+import{c}from"./index-DunbOMIr.js";/**
  * @license lucide-react v0.563.0 - ISC
  *
  * This source code is licensed under the ISC license.