+++ backend/gonzales/static/assets/video-BFm057oH.js
@@ -1,4 +1,4 @@
-import{c as e}from"./index-C_tx9rzG.js";/**
+import{c as e}from"./index-DunbOMIr.js";/**
  * @license lucide-react v0.563.0 - ISC
  *
  * This source code is licensed under the ISC license.